refactor-and-fix-openaichat | fix
