+++ llms/openai/openaillm_option.go
@@ -9,14 +9,6 @@ const (
 	organizationEnvVarName = "OPENAI_ORGANIZATION" //nolint:gosec
 )
 
-type APIType string
-
-const (
-	APITypeOpenAI  APIType = "OPEN_AI"
-	APITypeAzure   APIType = "AZURE"
-	APITypeAzureAD APIType = "AZURE_AD"
-)
-
 const (
 	DefaultAPIVersion = "2023-05-15"
 )
@@ -27,7 +19,7 @@ type options struct {
 	baseURL      string
 	organization string
 
-	apiType    APIType
+	apiType    openaiclient.APIType
 	apiVersion string // required when APIType is APITypeAzure or APITypeAzureAD
 
 	httpClient openaiclient.Doer
@@ -70,7 +62,7 @@ func WithOrganization(organization string) Option {
 
 // WithAPIType passes the api type to the client. If not set, the default value
 // is APITypeOpenAI.
-func WithAPIType(apiType APIType) Option {
+func WithAPIType(apiType openaiclient.APIType) Option {
 	return func(opts *options) {
 		opts.apiType = apiType
 	}
